pyproject.toml
[tool.poetry]
name = "baselinker_to_wix"
version = "0.1.0"
description = "Default description of the project!"
authors = ["Izbushko <izbushko@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
alembic = "^1.13.0"
fastapi = "^0.105.0"
websocket = "^0.2.1"
psycopg2-binary = "^2.9.9"
sqlmodel = "^0.0.14"
pydantic-settings = "^2.1.0"
passlib = "^1.7.4"
asyncpg = "^0.29.0"
uvicorn = "^0.24.0.post1"
pyjwt = "^2.8.0"
pydantic = {extras = ["email"], version = "^2.5.2"}
python-multipart = "^0.0.6"
gunicorn = "^21.2.0"
bcrypt = "4.0.1"
aiohttp = "^3.11.12"
celery = "^5.4.0"
redis = "^5.2.1"
requests = "^2.31.0"
beautifulsoup4 = "^4.13.3"
keepa = "^1.3.14"
httpx = "^0.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
